Hello relearners,
I don't know if I'm too late to bring this in :

Would it make sense to use some of the metafonts and stroke-fonts we created during the workshop for the layout?
There's a handy tool (mftrace) which converts .mf files into ps1 fonts using autotrace, which means we can translate all the stroke fonts into real fonts. I did a test with some Herschey fonts and I think the quality is good.

best,

Simon